+++ src/sequentially.ts
@@ -1,21 +1,5 @@
 import { isPromise } from './is_promise';
-import { First, Length, Tail } from './utilities';
-
-type HasPromise<
-  R extends ReadonlyArray<unknown>,
-  A extends boolean = false
-> = A extends true
-  ? A
-  : Length<R> extends 0
-  ? A
-  : HasPromise<Tail<R>, First<R> extends Promise<unknown> ? true : false>;
-
-type ExtractReturnTypes<
-  F extends ReadonlyArray<(...args: ReadonlyArray<unknown>) => unknown>,
-  R extends ReadonlyArray<unknown> = []
-> = Length<F> extends 0
-  ? R
-  : ExtractReturnTypes<Tail<F>, [...R, ReturnType<First<F>>]>;
+import { Tail, First, Length, HasPromise, ReturnTypesOf } from './utilities';
 
 type IsParametersEqual<
   F extends ReadonlyArray<(...values: ReadonlyArray<any>) => unknown>,
@@ -46,7 +30,7 @@ export const sequentially = <
   ...values: IsParametersEqual<V> extends true
     ? Parameters<First<V>>
     : ReadonlyArray<unknown>
-): HasPromise<ExtractReturnTypes<V>> extends true ? Promise<void> : void =>
+): HasPromise<ReturnTypesOf<V>> extends true ? Promise<void> : void =>
   fns.reduce(
     (result, fn) =>
       isPromise(result) ? result.then(() => fn(...values)) : fn(...values),